requirements.txt
Django==1.7.1
pymlconf
psycopg2
python-memcached
django-mbtiles
Pillow
requests
